backend/requirements.txt
fastapi
fastapi-mail==1.4.1
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pandas
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.4.0
email-validator
itsdangerous
opencv-python-headless==4.10.*
pillow==10.*
pytesseract==0.3.10
python-multipart==0.0.9
apscheduler==3.10.4
tzlocal>=5.0
